refactor: display 다루거나 validate 수행하는 함수들 모듈화
